refactor: move helper functions to helper pkg
